# JWTTokenAuthInAspNet6/ViewModels/LoginVM.cs
namespace JWTTokenAuthInAspNet6.ViewModels
{
    public class LoginVM
    {
        //public string? Id { get; set;}
        public string? UserName { get; set; }
        public string? Password { get; set; }
    }
}