# JWTTokenAuthInAspNet6/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "JWTToken_Auth_API",
    "version": "v1"
  },
  "paths": {
    "/api/Auth/Register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/RegisterVM"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Auth/Login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/LoginVM"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "LoginVM": {
      "type": "object",
      "properties": {
        "userName": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "RegisterVM": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "accountType": {
          "type": "string"
        },
        "phoneNo": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "shopName": {
          "type": "string"
        },
        "businessType": {
          "type": "string"
        },
        "userRole": {
          "type": "string"
        },
        "isDeleted": {
          "type": "boolean"
        }
      }
    }
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT Authorization header using the Bearer scheme. \r\n\r\n Enter 'Bearer' [space] and then your token in the text input below.\r\n\r\nExample: \"Bearer 1safsfsdfdfd\""
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}